Proof-of-concept local table access
In use cases where training is centralized (i.e. the trainer and
reverb server are launch from the same python process) and limited
compute is available, it is signficantly faster to access the table
directly instead of GRPC. This commit adds a new dataset and
sampler that demonstrate this idea with minimal changes to other
code.

A change required for this feature is changing the hash inside Table to
std::hash from absl::hash. absl::hash uses per process randomized seeds,
and table lookups do not work in the dataset.  std::hash typically
returns the value itself for integer types, which works well for this
use case since the keys are sufficiently randomized already.
In general, we do need to worry about using objects created in the
pybind module within a tf dataset, since absl is not designed to
support that well.
See https://github.com/abseil/abseil-cpp/issues/834.